hacks to generate *good* source maps
This patch ensures that the generated source maps include the entries
for "file" (the minimized/optimized source file) and "sources" (the
original files).

It does not work with JSTs; there may be other cases where it falls
down.

It remains to be seen whether this is a good idea for Jammit, or a good
starting point for a different project.